requirements.txt
asyncio
tqdm
openai
python-dotenv
black
tiktoken
aiohttp
pyyaml
pydantic
